Make flake8 plugin work with flake8 v2 and v3
They live in separate files; the correct one gets loaded based on flake8
version:

- `teamcity/flake8_v2_plugin.py`
- `teamcity/flake8_v3_plugin.py`
